fix some ui bugs
fix some ui bugs

+++ app/src/main/java/com/example/jesta/viewmodel/JestaDetailsViewModel.java
@@ -189,7 +189,7 @@ public class JestaDetailsViewModel extends ViewModel {
     }
 
     public void ownerFinishFavor(String transactionId, float rating, String comment) {
-        GraphqlRepository.getInstance().ownerFinishFavor(transactionId, (int) rating, comment);
+        GraphqlRepository.getInstance().ownerFinishFavor(transactionId, rating, comment);
         _navigationHelper.navigate(null);
     }
